PIM-3595: move saver and remover interfaces in storage utils component
