+++ spec/Pim/Bundle/CatalogBundle/Manager/AttributeGroupManagerSpec.php
@@ -22,17 +22,17 @@ class AttributeGroupManagerSpec extends ObjectBehavior
 
     function it_is_a_saver()
     {
-        $this->shouldImplement('Akeneo\Component\Persistence\SaverInterface');
+        $this->shouldImplement('Akeneo\Component\StorageUtils\Saver\SaverInterface');
     }
 
     function it_is_a_bulk_saver()
     {
-        $this->shouldImplement('Akeneo\Component\Persistence\BulkSaverInterface');
+        $this->shouldImplement('Akeneo\Component\StorageUtils\Saver\BulkSaverInterface');
     }
 
     function it_is_a_remover()
     {
-        $this->shouldImplement('Akeneo\Component\Persistence\RemoverInterface');
+        $this->shouldImplement('Akeneo\Component\StorageUtils\Remover\RemoverInterface');
     }
 
     function it_throws_an_exception_when_removing_an_attribute_group_and_the_default_group_does_not_exist(
