+++ spec/Pim/Bundle/CatalogBundle/Manager/AttributeManagerSpec.php
@@ -31,17 +31,17 @@ class AttributeManagerSpec extends ObjectBehavior
 
     function it_is_a_saver()
     {
-        $this->shouldImplement('Akeneo\Component\Persistence\SaverInterface');
+        $this->shouldImplement('Akeneo\Component\StorageUtils\Saver\SaverInterface');
     }
 
     function it_is_a_bulk_saver()
     {
-        $this->shouldImplement('Akeneo\Component\Persistence\BulkSaverInterface');
+        $this->shouldImplement('Akeneo\Component\StorageUtils\Saver\BulkSaverInterface');
     }
 
     function it_is_a_remover()
     {
-        $this->shouldImplement('Akeneo\Component\Persistence\RemoverInterface');
+        $this->shouldImplement('Akeneo\Component\StorageUtils\Remover\RemoverInterface');
     }
 
     function it_instantiates_an_attribute()
