+++ spec/Pim/Bundle/CatalogBundle/Manager/FamilyManagerSpec.php
@@ -32,12 +32,12 @@ class FamilyManagerSpec extends ObjectBehavior
 
     function it_is_a_saver()
     {
-        $this->shouldHaveType('Akeneo\Component\Persistence\SaverInterface');
+        $this->shouldHaveType('Akeneo\Component\StorageUtils\Saver\SaverInterface');
     }
 
     function it_is_a_remover()
     {
-        $this->shouldHaveType('Akeneo\Component\Persistence\RemoverInterface');
+        $this->shouldHaveType('Akeneo\Component\StorageUtils\Remover\RemoverInterface');
     }
 
     function it_provides_a_choice_list($userContext, $repository)
