+++ spec/Pim/Bundle/CatalogBundle/Manager/ProductManagerSpec.php
@@ -64,17 +64,17 @@ class ProductManagerSpec extends ObjectBehavior
 
     function it_is_a_saver()
     {
-        $this->shouldImplement('Akeneo\Component\Persistence\SaverInterface');
+        $this->shouldImplement('Akeneo\Component\StorageUtils\Saver\SaverInterface');
     }
 
     function it_is_a_bulk_saver()
     {
-        $this->shouldImplement('Akeneo\Component\Persistence\BulkSaverInterface');
+        $this->shouldImplement('Akeneo\Component\StorageUtils\Saver\BulkSaverInterface');
     }
 
     function it_is_a_remover()
     {
-        $this->shouldImplement('Akeneo\Component\Persistence\RemoverInterface');
+        $this->shouldImplement('Akeneo\Component\StorageUtils\Remover\RemoverInterface');
     }
 
     function it_has_a_product_repository(ProductRepositoryInterface $productRepository)
