+++ src/Akeneo/Component/StorageUtils/Remover/RemoverInterface.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Akeneo\Component\Persistence;
+namespace Akeneo\Component\StorageUtils\Remover;
 
 /**
- * Bulk remover interface, provides a minimal contract to delete many business objects
+ * Remover interface, provides a minimal contract to remove a single business object
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-interface BulkRemoverInterface
+interface RemoverInterface
 {
     /**
-     * Delete many objects
+     * Delete a single object
      *
-     * @param array $objects The objects to delete
+     * @param mixed $object  The object to delete
      * @param array $options The delete options
      *
      * @throws \InvalidArgumentException
      */
-    public function removeAll(array $objects, array $options = []);
+    public function remove($object, array $options = []);
 }
