+++ src/Akeneo/Component/StorageUtils/Saver/SaverInterface.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Akeneo\Component\Persistence;
+namespace Akeneo\Component\StorageUtils\Saver;
 
 /**
- * Bulk saver interface, provides a minimal contract to save many business objects
+ * Saver interface, provides a minimal contract to save a single business object
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-interface BulkSaverInterface
+interface SaverInterface
 {
     /**
-     * Save many objects
+     * Save a single object
      *
-     * @param array $objects The objects to save
+     * @param mixed $object  The object to save
      * @param array $options The saving options
      *
      * @throws \InvalidArgumentException
      */
-    public function saveAll(array $objects, array $options = []);
+    public function save($object, array $options = []);
 }
